requirements.txt
openai
streamlit
gradio
flask
pdfplumber
python-docx
python-dotenv